refactor: rename suite-common/earn-api to suite-common/earn-stablecoin-api
