+++ suite-common/earn-stablecoin-api/src/api/discovery.ts
