+++ suite-common/earn-stablecoin-api/src/api/health.ts
