+++ suite-common/earn-stablecoin-api/src/api/index.schemas.ts
