+++ suite-common/earn-stablecoin-api/src/api/index.ts
