+++ suite-common/earn-stablecoin-api/src/api/yield-data.ts
