+++ suite-common/earn-stablecoin-api/src/constants/vaults.ts
