+++ suite-common/earn-stablecoin-api/src/hooks/useAllYieldOpportunities.ts
