+++ suite-common/earn-stablecoin-api/src/hooks/useEnterYieldOpportunity.ts
