+++ suite-common/earn-stablecoin-api/src/hooks/useExitYieldOpportunity.ts
