+++ suite-common/earn-stablecoin-api/src/hooks/useGetYieldOpportunities.ts
