+++ suite-common/earn-stablecoin-api/src/httpClient.ts
