+++ suite-common/earn-stablecoin-api/src/index.ts
