+++ suite-common/earn-stablecoin-api/tsconfig.json
