+++ suite-common/wallet-config/tsconfig.json
@@ -2,7 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
-        { "path": "../earn-api" },
+        { "path": "../earn-stablecoin-api" },
         { "path": "../wallet-constants" },
         { "path": "../../packages/device-utils" },
         { "path": "../../packages/type-utils" },
