+++ suite-native/module-accounts-management/package.json
@@ -15,7 +15,7 @@
         "@react-navigation/native-stack": "7.12.0",
         "@reduxjs/toolkit": "2.11.2",
         "@suite-common/device": "workspace:*",
-        "@suite-common/earn-api": "workspace:*",
+        "@suite-common/earn-stablecoin-api": "workspace:*",
         "@suite-common/fiat-services": "workspace:*",
         "@suite-common/graph": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
