+++ suite-native/module-accounts-management/tsconfig.json
@@ -3,7 +3,9 @@
     "compilerOptions": { "outDir": "libDev" },
     "references": [
         { "path": "../../suite-common/device" },
-        { "path": "../../suite-common/earn-api" },
+        {
+            "path": "../../suite-common/earn-stablecoin-api"
+        },
         {
             "path": "../../suite-common/fiat-services"
         },
