+++ suite-native/module-earn/package.json
@@ -17,7 +17,7 @@
         "@reduxjs/toolkit": "2.11.2",
         "@shopify/flash-list": "2.3.0",
         "@suite-common/device": "workspace:*",
-        "@suite-common/earn-api": "workspace:*",
+        "@suite-common/earn-stablecoin-api": "workspace:*",
         "@suite-common/earn-staking-api": "workspace:*",
         "@suite-common/formatters": "workspace:*",
         "@suite-common/message-system": "workspace:*",
