+++ suite-native/module-earn/tsconfig.json
@@ -4,7 +4,9 @@
     "include": [".", "**/*.json"],
     "references": [
         { "path": "../../suite-common/device" },
-        { "path": "../../suite-common/earn-api" },
+        {
+            "path": "../../suite-common/earn-stablecoin-api"
+        },
         {
             "path": "../../suite-common/earn-staking-api"
         },
